limited auth to just Rise Vision domains
gracefully blocks non-authorized users from accessing the rest of the site. Since there is no server component to this, it must be done client side.
While a user could modify the js to get around this, they will not be able to access any of the metrics since Google will restrict access to them server side